src/main/java/com/verygoodbank/tes/web/exception/GlobalExceptionHandler.java
package com.verygoodbank.tes.web.exception;

import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.ExceptionHandler;
import org.springframework.web.bind.annotation.RestControllerAdvice;

@RestControllerAdvice
public class GlobalExceptionHandler {

    @ExceptionHandler(InvalidTradeDataException.class)
    public ResponseEntity<String> handleInvalidTradeDataException(InvalidTradeDataException ex) {
        return new ResponseEntity<>(ex.getMessage(), HttpStatus.BAD_REQUEST);
    }
    @ExceptionHandler(InvalidCsvFormatException.class)
    public ResponseEntity<String> handleInvalidCsvFormatException(InvalidCsvFormatException ex) {
        return new ResponseEntity<>(ex.getMessage(), HttpStatus.BAD_REQUEST);
    }

    @ExceptionHandler(InvalidPriceFormatException.class)
    public ResponseEntity<String> handleInvalidPriceFormatException(InvalidPriceFormatException ex) {
        return new ResponseEntity<>(ex.getMessage(), HttpStatus.BAD_REQUEST);
    }


}
